// src/components/StoryModal/types.ts
export interface CommonProps {
  images?: Array<string>;
  duration?: number | undefined;
  containerStyle?: ViewStyle;
  enableProgress?: boolean | undefined;
  imageStyle?: ImageStyle;
  barStyle?: BarStyleProps;
  maxVideoDuration?: number | undefined;
}

export interface BarStyleProps {
  barActiveColor?: string;
  barInActiveColor?: string;
  barHeight?: number;
}
// ...
export interface CallbackProps {
  userStories?: StoriesType | undefined;
  story?: StoryType[] | undefined;
  progressIndex: number;
  userStoryIndex?: number | undefined;
}

export interface StoryContainerProps extends CommonProps {
  stories: StoryType[];
  userStories?: StoriesType;
  visible?: boolean | undefined;
  index?: number | undefined;
  userStoryIndex?: number | undefined;
  storyIndex?: number | undefined;
  isShowReply?: boolean | undefined;
  renderHeaderComponent?: (callback: CallbackProps) => JSX.Element;
  renderFooterComponent?: (callback: CallbackProps) => JSX.Element;
  renderIndicatorComponent?: () => JSX.Element;
  userProfile?: UserProps | undefined;
  footerView?: FooterViewProps | undefined;
  onComplete?: (viewedStories?: Array<boolean[]>) => void;
  renderCustomView?: (callback: CallbackProps) => JSX.Element;
  backgroundColor?: string;
  style?: ViewStyle;
  progressIndex?: number | undefined;
  headerViewProps?: ViewProps;
  customViewProps?: ViewProps;
  footerViewProps?: ViewProps;
  progressViewProps?: ViewProps;
  storyContainerViewProps?: ViewProps;
  showSourceIndicator?: boolean;
  sourceIndicatorProps?: ActivityIndicatorProps;
  videoProps?: VideoProperties;
  onChangePosition?: (
    progressIndex: number,
    userStoryIndex?: number,
  ) => void | undefined;
  previousStory?: () => void | undefined;
  nextStory?: () => void;
  customViewStyle?: ViewStyle;
  headerStyle?: ViewStyle;
  footerStyle?: ViewStyle;
}

export interface UserProps {
  containerStyle?: ViewStyle | undefined;
  userImage?: ImageSourcePropType | undefined;
  userName?: string | undefined;
  userMessage?: string | undefined;
  imageArrow?: ImageSourcePropType | undefined;
  onImageClick?: Function | undefined;
  onClosePress?: Function | undefined;
  userImageProps?: ImageProps | undefined;
  closeIconProps?: ImageProps | undefined;
  userNameProps?: TextProps | undefined;
  userMessageProps?: TextProps | undefined;
  userImageStyle?: ImageStyle;
  rootStyle?: ViewStyle;
  userNameStyle?: TextStyle;
  userMessageStyle?: TextStyle;
  closeIconStyle?: ImageStyle;
  customCloseButton?: any;
}

export interface FooterViewProps {
  isShowReply: boolean | undefined;
  onReplyTextChange: Function | undefined;
  onReplyButtonClick: Function | undefined;
}

export type StoryType = {
  id: number;
  url?: string;
  thumbnail?: string;
  type?: string | 'image' | 'video' | 'text';
  duration?: number;
  isReadMore?: boolean;
  isViewed?: boolean;
  isPaused?: boolean;
  created?: string;
  storyId?: number;
  title?: string;
};

export type StoriesType = {
  username?: string;
  profile?: string;
  title?: string;
  id: number;
  stories: StoryType[];
};

export type StoryRef = {
  pause: (pause: boolean) => void;
  handleLongPress: (visibility: boolean) => void;
  viewedStories: boolean[];
};